Ellipse.random_point and RegularPolygon.reflect
A single random_point is now generated; there should
be no need for multiple attempts as long as the random
number is made rational. The docstring was also cleaned
up.

The few tests for the RegularPolygon.reflect method were
replaced with a single random test which should never fail
but will always be different.