Evaluate polylog(s, z) for s in [-1, 0, 1], without exp_polar
polylog(s, z) has simple formulas for s in -1, 0, 1. They should be
used by eval, just as it uses the formulas for z in -1, 0, 1 without
the user needing to expand_func. For example:
```
>>> polylog(2, S(1)/2)
-log(2)**2/2 + pi**2/12
>>> polylog(1, S(1)/2)
log(2)     # previously this returned unevaluated
```
One has to draw the line somewhere, and I left s = -2 and the rest
un-evaluated, subject to the user applying expand_func

Per the discussion at #13853 the expansion for s = -1 is simply
-log(1 - z), as it appears in Wikipedia, Wolfram, etc.