Fold expressions with Piecewise before trying to integrate them
Piecewise class has its own integration method, to which functions are
dispatched if they meet the condition `if isinstance(function, Piecewise)`
However, this condition fails if there is a sum (or other combination)
of piecewise functions, and then the integration is unlikely to succeed.
Folding into one piecewise allows the Piecewise integration method to
take over.

Example:
```
x, y = symbols('x y')
integrate(sin(x*y)*y, (x, 0, 1), (y, 0, 1))

```
does not completely evaluate because the first integration produces an
expression with Piecewise:
```
-y*Piecewise((0, Eq(y, 0)), (-cos(y)/y, True)) + \
 y*Piecewise((0, Eq(y, 0)), (-cos(2*y)/y, True))
```
Subsequent attempt at integration fails, even though it eventually results
in folding the piecewise (too late for it to be dispatched to the right
method): the output is
```
Integral(Piecewise((0, Eq(y, 0)), (-cos(y) + 1, True)), (y, 0, 1))
```
With this commit, the output is 1-sin(1). Added as a test.