Numeric check for pivot being zero in Bareiss determinant method
Fixes #13835

Currently, the default determinant computation method, Bareiss,
accepts as a pivot any entry `!= 0`. This leads to "nan" output when
the pivot is a symbolic expression that simplifies to zero.
The added check to `_find_pivot` tests it for `== 0` after
substituting random rational numbers for any variables.
This has no noticeable impact on speed, and prevents zero pivot
choices in practice.