Modifications to `periodicity`
`periodicity` earlier used  a check after splitting the given function
using `decompogen` to check that the function being tested for periodicity
is not same as the original function, to avoid infinite recursion. This
check equated both the functions. This, however, led to the funcction being
stuck if the decomposition of the functions were the same. Equating the
decomposition of the functions during the check solves this issue.

Example when this happened:
```
>>> from sympy import *
>>> x = Symbol('x')
>>> periodicity(gamma(S(1)/7 + 1/x))

```