Un-randomize the test of polylog values
Random numbers in test_polylog_values replaced by a fixed set of complex
numbers.

Explanation

I recently added a test to check that the values of polylog for special
values of s agree with the mpmath floating point evaluation. The test picked
random complex numbers and asserted Abs(difference) < 1e-15.

This turned out to be a bad idea because polylog has a pole at 1, so when
a random number is very close to 1, the absolute error of mpmath computation
may sometimes (rarely) be greater than 1e-15.

It is only fair that this random test failure occurred on my own PR.