Use verify_numerically in testing polylog expansions
Random test test_polylog_values, introduced in #13852, turned out to
fail in rare cases when a random point came to close to the pole at 1.
Also, it did not use the utility function verify_numerically which is
designed for this purpose. Both issues are corrected now. Finally, only
s = 0 and s = -1 are tested since these are the only values of s for
which polylog is auto expanded.