Evaluate the argument of polar numbers
arg(exp_polar(w)) now returns the imaginary part of w, instead of
being mostly un-evaluated. For example, arg(exp_polar(4*pi*I)) is
returned as 4*pi, consistent with the purpose of existence of
polar numbers: provide continuous evaluation of functions such as
log and arg.

The current behavior is
```
>>> arg(exp_polar(4*pi*I))
-I*log((re(exp_polar(4*I*pi)) + I*im(exp_polar(4*I*pi))) /
    sqrt(re(exp_polar(4*I*pi))**2 + im(exp_polar(4*I*pi))**2))
```
which is inconsistent with
```
>>> log(exp_polar(4*pi*I))
4*I*pi
```

For comparison