Expression should be NaN when Mul "cancels out infinity"
Fixes #13904

Mul combines powers with identical base, adding their exponents.
When this results in exponent of 0, the term is simply dropped, which
is wrong when the base is infinite. There is already a simple check
that suffices for oo/oo but does not recognize (x+oo)/(x+oo) as a
meaningless expression. A more careful check is added: canceling a
term where the base is an Add or Mul with an infinite argument results
in NaN.

The full analysis of base expression seems impractical for such a basic
method as Mul, which needs to be very fast. Many expressions containing oo
are in fact finite, like `Pow(2, -oo)` or `periodic_argument(I, oo)`.
The check for Add and Mul should be enough for common cases like substitution
in a rational function that presently results in an incorrect value:
```
var('a b')
r = (1/(a+b) + 1/(a-b))/(1/(a+b) - 1/(a-b))
r.subs(b, a)   # returns 1, contrary to any interpretation of the formula
```