Refer the computation of arg(polar number) to unbranched_argument
The argument, like logarithm, is unbranched on the Riemann surface
where polar numbers live. So it should be computed as unbranched
when the argument is a polar number.

Also, arg now uses .as_real_imag because using re and im just ends up
calling the same as_real_imag method twice, for real and imaginary
parts.

In the process I also organized the import statements.