Limit arg(polar number) eval to arg(exp_polar)
While it is tempting to send all polar numbers over to
unbranched_argument, the reality is that changing
arg(polar_lift(x)) to periodic_argument(x, oo) is not
really any simplification, and turns out to be
counterproductive for Meijer G integration.

So this commit limits arg evaluation to
arg(exp_polar(w)) = im(w) which is certainly simpler.