Correct handling of keyword arguments in Derivative.doit
Fixes #12996

Currently, the doit method of Derivative accepts keyword arguments, handles
two of them (evaluate and simplify) and tries to pass the rest to any new
derivatives created in the process, which raises an error because there are
no keyword arguments for `Expr.__new__`. This is corrected now.

The bug manifested itself in #12996 because a derivative was created in the
process of integration, and then received manual=False parameter (intended for
Integral) which it tried to pass to `Expr.__new__`.