Convert more Pow expressions to field elements
Fixes #13947

The `FracField._rebuild_expr` method did not try hard enough to express
a given Pow in terms of field generators: it just expected the exponent
to be integer, and the base be in the field. This missed, for example,
that `Z(2**pi)` contains `2**(-pi)`. Now the Pow and ExpBase instances
are treated more carefully, by comparing their bases to the bases of
Pow or ExpBase among generators, and seeking an integer ratio of exponents.