Evaluate limits of sequences with alternating signs
At present limit_seq is unable to calculate the limits of
sequences that have sign alternation, such as
`limit_seq((-1)**n / n, n)`. Now, when the expression
contains a negative term raised to a power involving n, the
computation of limit splits in two cases, of even and odd
n. If both subsequences x[2*n] and x[2*n+1] have the same
limit, then that is the limit of x[n].

This involved moving most of the former `limit_seq` to a helper
function `_limit_seq`. Additionally, the index is now declared
to be positive and integer, in case the user forgot to do so;
this may improve the performance of the algorithm.

Incidentally, I cleaned up the docstrings in limitseq.py which
had some typos and sometimes opaque explanations.