fix the imaginary number printing
As pointed out in the issue https://github.com/sympy/sympy/issues/13977, before this fix the program generates wrong octave code when it comes to imaginary numbers, e.g.```sqrt(3)i```. Now it is set to make multiplication explicit, i.e., sqrt(3)*i. It may not be the best way since i dont know this project, but it is the fastest that solved my personal issue.