Make geometric series summation more robust
This commit primarily targets issue #13979.  Now the pattern matching
for geometric series summand is turned into a more robust two-step
process, for each of which some shallow pre-processing is performed to
ensure the success of the matching when the summand can actually be
written into the form required by the pattern.