Symbolic Riemann sums for definite integrals
Fixes #12707

Adds a Boolean flag `evaluate` to `as_sum` method, True by default for
backward compatibility.

Also allows a symbolic number of subintervals to be used, which is both
natural and more interesting for a symbolic library: for some simple
functions the sum can be evaluated with a general n, showing that it
converges to the integral (and at different rates for different rules).

Concerning the values at endpoints: by the definition of a Riemann sum,
direct evaluation `f(a)` or `f(b)` is performed, not a limit as x->a or
x->b. For example,
```
Integral(sign(x), (x, 0, 1)).as_sum(1, 'trapezoid')
```
should be (f(0)+f(1))/2 = 1/2, not 1 as it was returned previously.

A link to Wikipedia is updated, as the previously referenced article was
merged.