Reorder Piecewise output of integration, so that generic answer is first
Fixes #7032

Currently, when `integrate` presents a solution as a Piecewise, it puts
an edge case first, for example

```
>>> var('x y')
>>> integrate(exp(x*y), x)
Piecewise((x, Eq(y, 0)), (exp(x*y)/y, True))
```

With this patch, the output is

```
Piecewise((exp(x*y)/y, Ne(y, 0)), (x, True))
```

An exception is made for Piecewise with multiple edge cases, such as
Eq(k, 0), Eq(m, 0), Eq(k, m), Eq(k, -m)... In such a situation, putting
the generic case would necessarily decorate it with a long condition,
and then negated pieces of that condition would be repeated again in
the edge cases. The result is a formula that is not only longer but
also harder for SymPy to work with.

Naturally, a lot of test cases had to be changed.