When integrating by parts, do not use u = non-poly algebraic
There is no gain in differentiating non-polynomial algebraic
expressions when integrating by parts, at least in the elementary
situations that manualintegrate deals with. Trying to do this leads to
infinite loops, Now this choice of u, dv is prevented.

Example: in current master `integrate((1+sin(x))/x, x)` gets stuck in
an infinite loop of this kind, as the exponent of x becomes more and
more negative under differentiation. Now this integral gets evaluated,
because after manualintegrate fails, it goes back to meijerint as two
terms 1/x and sin(x)/x, resulting in log(x) + Si(x).

There was already a check to prevent this when the algebraic is
multiplied by exp, but this was overly specific.