If the term of a Product is Add, just factor it
Fixes #14036

The product-evaluating code handling the case term.is_Add was trying
questionable things, such as rewriting as sum of logarithms (only
possible with positive terms), or computing the product of denominators
and then deciding that the rewrite didn't work if the result was
is_Number (when it fact it could mean the product of computed).

Realistically, all we can do with an Add in a product is to factor it,
and let the code for term.is_Mul do the rest. If it doesn't factor, give
up. This is what is done now.