Don't trade one logarithm for another when integrating by parts
In current master (and in 1.1.1) `integrate(log(1/x)/(1+x), x)` and
similar integrals hang because manualintegrate keeps applying
integration by parts. This is now prevented.

Aside: Meijer G can handle this integral if x is declared positive,
because then log(1/x) can be rewritten to -log(x). But it fails without
this assumption.