Use different tractable forms of erf for different points
Addresses a part of #13750 but does not solve it.

Presently, every time a limit involving erf is taken, erf is rewritten
as 1 - (something small as x->oo), which operates under assumption that
the argument of erf tends to oo. As noted in #13750, this leads to wrong
limit values: e.g., limit(erf(sqrt(x) - x), x, oo) is claimed to be 1
(it's -1). Also, very basic limits at finite points, such as
limit(erf(x)/x, x, 0), return unevaluated. Both of these are corrected
by this patch, which changes erf.eval_rewrite_as_tractable so it depends
on the limit of the argument of erf.

This is only a partial solution, because it only works if the argument
of erf has 1 symbol. With multiple symbols, the rewriting method has no
knowledge of which symbol tends to oo. The possibility of passing this
information to _eval_rewrite methods should be considered.