14052: Piecewise._intervals better handles Eq
Any intervals of zero width are ignored; this also removes
the incorrect identification of a condition that is true
at a given value of sym as being a default condition.

And(x<2, Eq(x, 1)) is true, but this doesn't mean that
the domain in which it is true is (-oo, oo), only that
it is true at x == 1. Normally, such a condition would
be handled by instantiation:

>>> Piecewise((1,And(x<2, Eq(x, 1))))
Piecewise((1, Eq(x, 1)))

But such conditions involving And and Eq are generated
when solving conditions involving inequalities, e.g.

>>> solve(sin(x)<=0)
Eq(x, 0) | ((pi <= x) & (x <= 2*pi))