More general unpolarification for polylog argument
A follow-up to #13965. Closes #14075.

In #13965, the argument of polylog was unpolarified if it has is_polar flag
and was at most 1 in absolute value. It turned out that expressions like
-exp_polar(I*pi) have is_polar = False. This behavior seems intended, see
https://github.com/sympy/sympy/blob/master/sympy/core/mul.py#L1174
and the test
https://github.com/sympy/sympy/blob/master/sympy/core/tests/test_arit.py#L1714
which asserts that if p is polar, -2*p is not polar.

I can understand the logic of this approach to polarity, since
multiplying a polar number by a non-positive complex number is not a
well-defined operation on the Riemann surface of the logarithm. But for
the purpose of evaluating polylog all that matters is that we have some
exp_polar and want to get rid of it. So rather than try redefining the
behavior of is_polar, I changed the condition from z.is_polar to
z.has(exp_polar)