In log_to_real, pair conjugate roots without losing any pairs
This implements the solution [proposed by
@jksuom](https://github.com/sympy/sympy/issues/12050#issuecomment-339047735).
Presently, when `log_to_real` converts complex logarithms to atan, it
only picks roots with positive imaginary part, to avoid duplication.
This is a problem because it's not always known which of two roots has
positive imaginary part. In the latter case, we should just pick one
instead of ignoring the pair of roots completely.