Add repeated_decimal as a transformer on the standard library tokenize
Unfortunately, there is currently a regression, which is that .[1] doesn't
work. To handle this, we need to detect when . is an OP that cannot apply to
whatever is before it. Alternately, we could always parse . (as OP) + [INT] as
a repeated decimal, since it would never be valid syntax anyway.