Fix repeated decimals tokenizing when the repetend starts with 0
Python 2 tokenizes the first non-89 digits separately, and Python 3 tokenizes
the first 0 digits separately.

This has a minor regresssion, which is that sympify('.[0 1]') now parses the
same as sympify('.[01]'), because the tokenizer does not distinguish between
the two. This could be fixed by checking the column positions, but it doesn't
seem worth it.

Also:

- Checks for x are not necessary, because hex numbers cannot be floats.

- Fix some indentation.

- Make sure to clear num after it is added to result.