Check for logarithmic singularities in _eval_interval
After an antiderivative is found, expr._eval_interval is supposed to
evaluate it on the interval [a,b], paying attention to singularities. It
currently detects only rational singularities, not logarithmic ones,
which are common in integrals of 1/x, tan(x), etc.

A check for logarithmic singularities is added. Along the way, the check
that singularities are indeed within the interval is fortified against
TypeErrors coming from relationals of unknown value (typical for
multiple integrals).