This PR closes issue #14013.
Sum(sin(n), (n, 1, oo)).is_convergent() now returns S.false.
The limit of sin(n) for n tending to infinity is not zero and so the series
is not convergent.
In such situation, the limit of sin(n) returns an AccumulationBounds.
Thus a check for AccumulationBounds was added.

A test for the issue is added.