Unpolarify pFq with p == q+1 in the unit disk
Currently the argument of pFq is unpolarified if p <= q, when the
function is entire. However, when p == q+1 we still have a convergent
power series in the unit disk, and the results of SymPy computations
refer to the branch determined by this power series. Hence, the argument
can be unpolarified in this case, provided it is <=1 in absolute value.