13098: resolve expr-int(expr) in floor
A check is made that enough precision is used
to calculate the expr in case it is large. But
extra precision must also be used to calculate
the expr in case it is very close to an integer
or else the expression will be computed as the
integer. This is done by computing the exponent
of expr - int(expr) (which should be negative
since the difference should be less than 1) and
then subtracting it from 2. So if the exponent
were -70 the minimal precision needed to capture
the leading digit would be 70 + 1 (in case of
rounding) + 1 (for safety) = 72.  Is that enough
safety?