fix ``print_gtk()`` in python3
``print_gtk()`` doesn't work in python3.
Calling it will raise a exception::

	Traceback (most recent call last):
      ...
      File ".../sympy/printing/gtk.py", line 16, in print_gtk
        file.write( c2p(mathml(x), simple=True) )
    TypeError: a bytes-like object is required, not 'str'

It's because the temp file is opened with the mode 'wb',
which accepts ``str`` in python2, but ``bytes`` in python3.

I fixed it by using mode 'w' which works in both python2 and python3.
I also used ``subprocess`` to replace ``os.system()``,
which avoids hacking like space in file path.