Abs of integer powers can be applied to the base
The current implementation of Abs(Pow(base, exponent)) evaluates it only
if base is real. But when the exponent is an integer, it is always true
that we get Pow(Abs(base), exponent). For example, `Abs(I**n)` now
evaluates to 1 when n is an integer symbol. In current master it returns
`sqrt(I**n*(-I)**n)` which isn't as easy to SymPy algorithms (in
particular, `limit_seq`) to deal with.