Better dispatch of functions in ring series module
Currently, rs_series function (which expands an expression in a power
series) dispatches a function func to rs_func based on str(func) being
in the _convert_func dictionary. This has several drawbacks:
(a) The dictionary has to be updated every time a new rs_func is
added, and this has not happened for several of the already implemented
functions;
(b) If a user defines own function with matching name, the expansion will
be applied to it.
(c) If a function is not found in the dictionary, an unhandled KeyError is
thrown.

All are corrected here. The dictionary is replaced by a lookup for "rs_"
+ func.__name__ in globals(). (It would be better if these were methods
of some object, but this would require refactoring and deprecation.)
Undefined functions, as well as the functions for which no rs_func
exists, raise NotImplementedError.