Limits should not be sequential by default
Currently, if heuristics and gruntz both fail, the limit at infinity is
considered as the limit of a sequence by default: `if
hints.get('sequence', True) and z0 is S.Infinity:`. This leads to wrong
answers, such as
```
var('x')
limit(exp(I*x)*sin(pi*x), x, oo)  # returns 0
```
To correct this, the keywords arguments of `Limit.doit()` and of `limit`
are made explicit. The default value of `sequence` is changed to None,
which means to infer the limit type from the variable being integer. The
tests in `test_limit_seq`, which relied on the previous behavior, are
changed to use `sequence=True`. A test is added to ensure that
`limit(exp(I*x)*sin(pi*x), x, oo)` now raises NotImplementedError.