function: Added the differentiability test in Subs class method doit()
In Subs class the doit() method earlier used to return values even
at points where the derivative of the function doesn't exists.
After making this commit this should not happen for single variable
functions. However, the work on multi-variable functions is in progress
and soon, a PR on that will be made.