Avoid arg(0) < pi error by checking the argument of arg first
In two places in meijerint the condition |arg(1-z)| < pi needs to be
checked. The direct evaluation is problematic because if z is 1, the
left side is NaN, and NaN < pi is a TypeError. For this reason, z == 1
has to be checked first in a helper function. And this was already done,
but only in one of two places.

Specifically, the variablez zos and zso are checked in the above way;
previously only zso was. The helper function `_cond` is changed to take
a parameter, and its logic is simplified to `return False if z == 1 else
abs(arg_(1 - z)) < pi`.