Remove sequence flag from Limit.doit()
Currently, if heuristics and gruntz both fail, the limit at infinity is
considered as the limit of a sequence by default: `if
hints.get('sequence', True) and z0 is S.Infinity`. This leads to wrong
answers, such as in the linked issue.

Per the discussion in #14327, the undocumented `sequence` parameter of
`limit` is removed. The docstring of `limit` now refers to `limit_seq`
for computing limits of sequences. The remaining keywords arguments of
`Limit.doit()` and `limit` are made explicit. Some tests in
`test_limits`, which relied on the sequential behavior, are moved to
`test_limitseq`. A test for `limit(exp(I*x)*sin(pi*x), x, oo)` is added.