The values of transcendental functions at complex infinity
Added or corrected the values of exponential, trigonometric, and
hyperbolic functions (with inverses) at complex infinity.

I noticed the inconsistency in that `sin(zoo)` is unevaluated but
`re(sin(zoo))` is NaN. Indeed, `sin(zoo).rewrite(exp)` is NaN. I think
that for any function that does not have a (finite or infinite) limit at
complex infinity, eval should return NaN when the argument is zoo
(unless AccumBounds can be used for more specific result). The current
behavior of exp and trigonometric functions at zoo:

- exp, sin, cos, tan, cot, sec, csc, asin, atan, acot return unevaluated
- acos returns zoo   (correct)
- asec returns pi/2  (correct, since it is acos(1/z))
- acsc returns 0     (correct, since it is asin(1/z)

For the previously unevaluated functions, the following should be the values
at zoo:

- exp, sin, cos, tan, cot, sec, csc should return NaN
- asin should return zoo, as acos does.
- atan should return AccumBounds(-pi/2, pi/2) as its imaginary part tends to 0
while the real part is between these bounds.
- acot should return 0, being atan(1/z).

For the hyperbolic trigonometric functions and their inverses:

- sinh, cosh, tanh, coth, sech, csch presently return NaN at zoo (correct)
- asinh returns zoo, correct
- acosh returns oo; incorrect. Should be zoo as for asinh. There is also a test asserting acosh(I*oo) == oo and acosh(-I*oo) == oo, but the actual
limits of acosh(I*y) are oo+I*pi/2 at oo and oo-I*pi/2 at -oo.
- atanh returns nan; correct but `I*AccumBounds(-pi/2, pi/2)` would be more
accurate
- acoth and acsch return 0; correct
- asech returns NaN; correct but `I*AccumBounds(-pi/2, pi/2)` would be more
accurate. (Note the difference from asec: although asech(z) = acosh(1/z),
we don't get a limit because the principal branch of acosh is discontinuous
at 0.)

Side effect: when testing, I noticed that `Expr._eval_interval` does not
pay attention to AccumBounds in the output. Added them to the list of
things that necessitate taking a limit instead of direct substitution.