primitive_element: improvement
Currently, primitive_element (non-experimental version) does not
correctly handle sequences of input extensions that are not
linearly disjoint because the minimal polynomial (over Q) of one
extension may be reducible over another extension. (See
https://github.com/sympy/sympy/issues/14117#issuecomment-364207683)

Here, the minimal polynomial of each incrementally constructed
extension is replaced by a factor that is irreducible over the
next input extension.

Also, the polynomials whose sqf_norm is computed are chosen so that
the coefficients s need not be cumulatively multiplied.

Finally, input extension generators are not converted to AlgebraicNumbers
as it seems unnecessary.