Disable simplification of inverses unless a flag is set
Currently, `simplify` cancels inverses without considering whether this
is valid: `simplify(asin(sin(x)))` returns `x` which is a different
thing, for example, if x is pi. Also, this cancellation is implemented
in such a way that `simplify(2*asin(sin(x)))` returns unchanged,
breaking consistency.

Recognizing that sometimes the above simplification is desired, a new
function `inversecombine` is added, which performs the cancellations of
the above kind. A new flag `inverse` is added to `simplify`; if set to
True, then `inversecombine` will be called. The default is False.