Improve handling of MatrixSymbols in gcd_terms and trigsimp
MatrixSymbol objects sometimes cause trouble because one of their args
is a string, which is unusual for SymPy.
```
>>> M = MatrixSymbol('M', 3, 3)
>>> M.args
('M', 3, 3)
```
The helper function `handle` of `gcd_terms` applies itself recursively
to non-Expr objects `a` as follows:
```
type(a)([handle(i) for i in a])
```
This is a problem when a is a string, since `[i for i in a]` is a list
of string; so the function gets applied to the first character again.
This is fixed now.

Issue #14421 also exposed a defect in trigsimp:
```
trigsimp(sin(t)**2*V[1,0] + cos(t)**2*V[1,0])  # no simplification
```
The reason was an unnecessarily recursive check in `_is_Expr` part of
`_eapply` helper. This check was intended to ensure that an object and
all of its args are instances of Expr. But in fact it demanded also args
of args, etc, to be Expr. Thus, any expression containing MatrixSymbol
failed the check. The recursive behavior was unnecessary, because the
functions used by `_eapply` do not recurse through expression tree.
Checking that the first-level arguments are Expr is sufficient, and this
is what is done now.