Implement Cholesky and LDL decompositions for Hermitian matrices
Currently, the `cholesky` method checks whether the matrix is symmetric.
This is wrong in two ways: complex symmetric matrices get through (and
they shouldn't), while complex Hermitian matrices do not (and they
should). The check is adjusted, and the implementation is extended to
Hermitian matrices.

Additionally, a ValueError is raised if the matrix is found to not be
positive definite. Previously there was no check for that.

Similar adjustments are made to `cholesky_solve`, `LDLdecomposition`,
and `LDLsolve`. Test are added.