Integrals: special treatment for substitutions u=(a*x+b)**(1/n)
Substitutions `u = (a*x+b)**(1/n)`, with integer n, are useful because x
is a simple expression in u: `x = (u**n - b)/a`. This makes it
reasonable to use this inverse to perform substitution in
manualintegrate. The general check for substitutions, namely that
`(f/u').subs(u, Dummy)` does not have the original variable of
integration, would not necessarily apply here.