Non-recursive Poisson sampling
The current sampling algorithm for the Poisson distribution uses the
strategy of doubling n (the sample value), which cannot work when the
sample to return is 0. Now the case of 0 is dispatched before entering
the doubling loop. Also, the binary search is made non-recursive to
avoid potential stack overflow issues when the search needs to take
long.