numbers: modified an error line in Integer
On giving input as `Integer(oo)`, it produced an error that `Integer can
only work with integer expressions`, which is a bit inaccurate as
Integer also works on input such as log(10), exp(3) and more inputs
which can be rounded off to an integer type. So I have modified the
error line.
@jksuom ping please, is the modified message up to the mark?