Exclude zero roots of resultant in the Descartes method for quartics
The method `_roots_quartic_euler` picks the largest rational root of the
resolvent `64*x**3 + 32*p*x**2 + (4*p**2 - 16*r)*x - q**2` expecting it
to be positive because the constant term is `-q**2`. This expectation is
unwarranted since the coefficients may be complex, or q could be zero.

Instead, the zero root(s) should be explicitly filtered out. Taking
maximum is fine after that.