give x0 and x1 for initial guess
The default x1 value is x0 + .25. That's probably why the
convergence was so bad: if the interval was smaller than .25,
the x1 value was set outside of the interval. Now the 0.25 is
scaled to the size of the interval and an explicit value of
x1 is given.

Here is another method that works but unless the degree of the
polynomial and the precision is low -- the product of the two--
then it can take a long time to compute all of the roots. So
even though assignment is fast, computing all roots is slow.

def _eval_evalf(self, prec):
    roots = self.poly.nroots(prec)
    i = self._get_interval()
    if self.is_real:
        a, b = map(lambda i: Rational(str(i)), (i.a, i.b))
        for r in roots:
            if a <= r <= b:
                return r
    ax, bx = i.ax, i.bx
    ay, by = i.ay, i.by
    ax, bx, ay, by = map(lambda i: Rational(str(i)), (ax, bx, ay, by))
    for r in roots:
        X, Y = r.as_real_imag()
        if ax <= X <= bx and ay <= Y <= by:
            return r
    raise NotImplementedError('method did not find corresponding root')