CartanType_generator: Correct instantiation
The current implementation of liealgebras has two ways to instantiate a
CartanType and hence also a RootSystem. For example,
c = CartanType('A7')
and
c = CartanType(['A', 7])
give the same thing. However, the former method only reads the first digit,
so, for example, CartanType('A34') will be exactly the same as CartanType('A3')
although it should be CartanType(['A', 34]).

I corrected this mistake by explicitly verifying if the argument is a list or
a string, and acting accordingly.