an improved nc_simplify
This is a new version of nc_simpliy. Now inverses are handled so
that nc_simplify(a*b*(b**-1*a**-1)**2) = (a*b)**-1 and
nc_simplify(c**-1*b**-1*a**-1) = (a*b**c)**-1. Negative powers
are generally handled better, including this sort of thing:
nc_simplify(a*b**3*a*b**3*a*b) = (a*b**3)**3*b**-2. Additionally,
simplifications are now chosen on the basis of what will leave
the fewest arguments rather than the earliest and longest possible
repeated subterm as before. The repetitions themselves are found
by computing a matrix of overlaps of prefixes and suffixes of the
word instead of repeated scanning the word for each possible
repeated subterm length.