lambdify: Updated to support using moduels="autograd.numpy"
autograd (https://github.com/HIPS/autograd) provides methods
for differentiating (almost) arbitrary numpy code.  However,
it requires functions to be implemented with it's own thinly
wrapped version of numpy (autograd.numpy).  This patch adds
support and tests for lambdifying with this version of numpy
and thus allows one to use autograd on functions that are
lambdified with SymPy.

The use case for this feature is where you have autograd.numpy
code mixed in with functions lambdified by SymPy.  It is
possible to provide gradients to autograd that are lambdified
from sympy's symbolic diff, however, computing many gradients
or higher order gradients would require the user to supply more
and more lambdified symbolic gradients.