Add tests, both in docstring and separate file
There are 3 examples in the docstring of zb_recur. We check the
entire result of 2 and just the recurrence of 1. In a different file
we have a number of tests, for convenience we just check that
the function G/F returned by the function is correct. It is highly
unlikely this function be correct and the recurrence wrong.