Add function to compute sums via zb_recur
This function attempts to compute tricky definite sums by find a
recursion with zb_recur, this function presently has some issues
and is not doing as much as it could be. Nonetheless it allows one
to compute sums which sympy is seemingly unable to do (e.g. both
examples in the docstring).

An example,

F = k * binomial(2 * n + 1, 2 * k + 1), calling
zb_sum(F, n, 1, n, k) gives

4*2**(2*n - 3)*factorial(n - 1/2)/factorial(n - 3/2)

which is correct, and simplifies to 4^(n-1) * (2n - 1).

Presently zb_sum has some problems when it comes to actually making
use of the recurrence (which is working fine), for example rsolve
does not produce the correct result for the recurrence that
(-1)^k * binomial(2n, k)^3 satisfies and as a result can not arrive at
the sum.

Other changes -

Now zb_recur gives a value error rather than returning a string.

In zb_recur we at some point create a polynomial in k and we are
able to just provide the coefficients in a list, so we do that.

The docstring of zb_recur was slightly changed. Previously it
asserted that something it returned was a polynomial, while this
can be made to be true. There is no make it be so, the docstring
now better reflects reality.